Function sui_adapter::adapter::execute

pub fn execute<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error = E> + Storage>(
    vm: &MoveVM, 
    state_view: &mut S, 
    module_id: ModuleId, 
    function: &Identifier, 
    type_args: Vec<TypeTag>, 
    args: Vec<CallArg>, 
    gas_status: &mut SuiGasStatus<'_>, 
    ctx: &mut TxContext
) -> SuiResult<()>

Execute module::function<type_args>(object_args ++ pure_args) as a call from sender with the given gas_budget. Execution will read from/write to the store in state_view. IMPORTANT NOTES on the return value: The return value is a two-layer SuiResult. The outer layer indicates whether a system error has occurred (i.e. issues with the sui system, not with user transaction). As long as there are no system issues we return Ok(SuiResult). The inner SuiResult indicates the execution result. If execution failed, we return Ok(Err), otherwise we return Ok(Ok). TODO: Do we really need the two layers?