Function sui_adapter::adapter::generate_package_id

pub fn generate_package_id(
    modules: &mut [CompiledModule], 
    ctx: &mut TxContext
) -> Result<ObjectID, SuiError>

Given a list of modules, use ctx to generate a fresh ID for the new packages. If is_framework is true, then the modules can have arbitrary user-defined address, otherwise their addresses must be 0. Mutate each module’s self ID to the appropriate fresh ID and update its module handle tables to reflect the new ID’s of its dependencies. Returns the newly created package ID.