Function sui_adapter::adapter::resolve_and_type_check

pub fn resolve_and_type_check<Mode: ExecutionMode>(
    objects: &BTreeMap<ObjectID, impl Borrow<Object>>,
    module: &CompiledModule,
    function: &Identifier,
    type_args: &[TypeTag],
    args: Vec<CallArg>,
    is_genesis: bool
) -> Result<TypeCheckSuccess, ExecutionError>

Check that package_object, module and function are valid

Check that the the signature of function is well-typed w.r.t type_args, object_args, and pure_args

Return the ID of the resolved module, a vector of BCS encoded arguments to pass to the VM, and a partitioning of the input objects into objects passed by value vs by mutable reference