Function sui_adapter::adapter::verify_and_link

pub fn verify_and_link<E: Debug, S: ResourceResolver<Error = E> + ModuleResolver<Error = E> + Storage>(
    state_view: &S, 
    modules: &[CompiledModule], 
    package_id: ObjectID, 
    natives: NativeFunctionTable, 
    gas_status: &mut SuiGasStatus<'_>
) -> Result<MoveVM, SuiError>

Given a list of modules, links each module against its dependencies and runs each module with both the Move VM verifier and the Sui verifier.