Trait sui_adapter::execution_mode::ExecutionMode

pub trait ExecutionMode {
    type ExecutionResult;
    type ExecutionResults;

    fn allow_arbitrary_function_calls() -> bool;
    fn make_result<S: MoveResolver>(
        session: &Session<'_, '_, S>,
        module_id: &ModuleId,
        function: &Identifier,
        type_arguments: &[TypeTag],
        return_values: &SerializedReturnValues
    ) -> Result<Self::ExecutionResult, ExecutionError>;
    fn empty_results() -> Self::ExecutionResults;
    fn add_result(
        results: &mut Self::ExecutionResults,
        idx: TransactionIndex,
        result: Self::ExecutionResult
    );
}

Required Associated Types

type ExecutionResult

the type of a single Move call execution

type ExecutionResults

the gathered results from batched executions

Required Methods

fn allow_arbitrary_function_calls() -> bool

Controls two things:

the calling of arbitrary Move functions

the ability to instantiate any Move function parameter with a Pure call arg. In other words, you can instantiate any struct or object or other value with its BCS bytes.

fn make_result<S: MoveResolver>(
session: &Session<'_, '_, S>,
module_id: &ModuleId,
function: &Identifier,
type_arguments: &[TypeTag],
return_values: &SerializedReturnValues
) -> Result<Self::ExecutionResult, ExecutionError>

fn empty_results() -> Self::ExecutionResults

fn add_result(
results: &mut Self::ExecutionResults,
idx: TransactionIndex,
result: Self::ExecutionResult
)