Module sui_adapter::adapter

Structs

MoveVM

Enums

CheckCallArg

Small enum used to type check function calls for external tools. ObjVec does not exist for Programmable Transactions, but it effectively does via the command MakeMoveVec

Functions

generate_package_id

Given a list of modules, use ctx to generate a fresh ID for the new packages. Mutate each module’s self ID (which must be 0) to the appropriate fresh ID and update its module handle tables to reflect the new ID’s of its dependencies. Returns the newly created package ID.

missing_unwrapped_msg

new_move_vm

new_session

resolve_and_type_check

Check that package_object, module and function are validCheck that the the signature of function is well-typed w.r.t type_args, object_args, and pure_argsReturn the ID of the resolved module, a vector of BCS encoded arguments to pass to the VM, and a partitioning of the input objects into objects passed by value vs by mutable reference

substitute_package_id